Revert "Set Up Implementation for Osu UI and more"
This reverts commit 62a78cfee962ea77d8868260178e175294ce7ad6.
I did something wrong and would like this submitted as a merge request,
since it has some interesting design philosophy choices? I'll try to add
it in an intuitive way
