--- Assets/Models/Materials.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 608a18685e4bd4ac5807b1274ebacf6f
-folderAsset: yes
-timeCreated: 1489437545
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
