--- Assets/Models/Materials/defaultMat.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: cde3ce1ff71fc4b08bb6424614714cdc
-timeCreated: 1489437545
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
