--- Assets/Scripts/OsuMechanic.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 0b77eab69224a4740b36f4b977afdf59
-timeCreated: 1489440498
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
