+++ Assets/Scripts/Player.cs
@@ -8,6 +8,7 @@ public class Player : MonoBehaviour {
 	public float speed = 5.0f;
 	public float arrowSpeed = 40f;
 	public float arrowDmg = 5f;
+	public SpellController spellUI;
 
 	private Camera mainCam;
 	private GameObject cursor;
@@ -15,8 +16,6 @@ public class Player : MonoBehaviour {
 	private RaycastHit camRayHit; //Hit point of raycast
 	private Vector3 deltamovement;
     private GameObject gamecontroller;
-
-	private SpellAndArrowManager manager;
 
 	private double timeHeld;
 	private double test;
@@ -28,7 +27,7 @@ public class Player : MonoBehaviour {
 		gamecontroller = GameObject.FindGameObjectWithTag("GameController");
 		cursor = GameObject.FindGameObjectWithTag("Cursor");
 		mainCam = GameObject.FindGameObjectWithTag ("MainCamera").GetComponent<Camera>();
-		manager.setup ();
+		spellUI.setup ();
 
 		timeHeld = 0.0;
     }
@@ -51,23 +50,24 @@ public class Player : MonoBehaviour {
 		if (Input.GetMouseButtonUp (1))
 			rightClicked = false;
 
-		//Must pass this test first to see if we're firing a regular arrow or a basic arrow.
-		if (timeHeld < 1.5)
-			AimPlayer ();
-		else
-			//For some reason, this causes an error right here. Any way to fix it?
-			manager.toOsu ();
-
 		//Handles whether we're starting a spell or a regular arrow
 		if (rightClicked)
 			timeHeld += Time.deltaTime;
 		else {
 			timeHeld = 0;
 
-			//Spawns Basic Arrow. FIX SYNTAX
-			//GameObject toSpawn = Instantiate (manager.basicArrow, player.transform.position, player.transform.rotation, this.transform);
+			//We're just firing a regular arrow. CREATE OBJECT HERE
 		}
     }
+
+	void FixedUpdate() {
+		if (timeHeld < 1.0)
+			AimPlayer ();
+		else {
+			spellUI.firing = true;
+			SpellManager ();
+		}
+	}
 
 	void AimPlayer() {
 		camRay = mainCam.ScreenPointToRay (Input.mousePosition);
@@ -79,5 +79,13 @@ public class Player : MonoBehaviour {
 			transform.LookAt(targetPos);
 			transform.eulerAngles = transform.eulerAngles + 180 * Vector3.up;
 		}
+	}
+
+	void SpellManager () {
+		if (spellUI.firing) {
+			spellUI.Osu ();
+		}
+		else
+			timeHeld = 0;
 	}
 }
